Moving updates server data and sends a message back to the client.
Turns out if you just pass a struct, it passes the DATA of the struct.
By value. So I got to unwind that four or five layers, and now we're
passing pointers everywhere, which I'd hoped to avoid.

So much for that. I'm sorry, but this is the only way I could find.