Added a bullet-update thread.
Each time this thread wakes up (by default, once every 500ms),
it pushes a 'U' (for update) event into the queue.

When this event is popped, update_bullets() executes, which
currently handles moving bullets and deleting ones that go off
the edges of the screen.

It does not yet handle collision. I'm working on that.